feat: complete Bombing War game — 40 Dart source files (~3850 lines)
Game components:
- Aircraft: Interceptor (barrel roll), Heavy Bomber (armor), Stealth X-26 (cloak)
- Enemies: Infantry, RPG unit, pop-up Bunker, Underground Factory (boss)
- Projectiles: Bullet, Missile (homing), Bomb (gravity + AoE)
- Effects: Explosion particles, smoke trail

Game systems:
- CollisionSystem: circle-overlap detection, centralized
- ScoreSystem: combo multiplier (x2 after 3 kills/5s, x3 after 5 kills)
- ThreatSystem: radar detection bar, missile barrage at 100%
- WaveSystem: procedural enemy wave spawning

UI screens:
- MainMenu, Hangar (aircraft unlock), GameScreen, LevelComplete, GameOver
- Virtual joystick + weapon buttons overlay
- HUD: score, health, threat bar, weapon selector

Architecture:
- Template method pattern in AircraftComponent for aircraft specialization
- Data-driven infinite level generation (seed = level * 31337)
- SharedPreferences save/load for progress, coins, unlocked aircraft

+++ lib/config/game_config.dart
@@ -80,7 +80,7 @@ class GameConfig {
   static const int bunkerScore = 100;
   static const double bunkerSize = 36.0;
   static const double bunkerOpenDuration = 1.5;
-  static const double bunkerSalvoCount = 3.0;
+  static const int bunkerSalvoCount = 3;
   static const double bunkerSalvoCooldown = 5.0;
   static const double bunkerMissileDamage = 45.0;
